Change Universes from owning to borrowing Levels.
This allows us to remove all the clone()s, which hopefully speeds things
up since we don't need keep incrementing refcounts and references cost
nothing to drop in data structures that own them.

Possibly even nicer for cache would be storing the Level inline except
for a reference to the Dp... but we can deal with that later.