Adding undo capabilities to universes.
Looking ahead a bit, most of the time we'll have to roll back changes to
the universe context at some point during module checking.  There are a
few ways of trying to deal with this, but a simple one that works nicely
with the extremely imperative approach we've taken is to keep track of
an undo log (transactional database style).  Whenever you want to create
a new rollback segment, you can just start a new undo log that tracks
all the changes after that point (until the next segment).  When you
want to roll back, you just roll back all the existing segments in the
opposite order of their creation.  We make sure each operation is
deterministic and linearizable to ensure that this process is
straightforward (though it will not work correctly if you don't
precisely mirror the order of undos with the order things were added to
the log).

There's a lot of room for improvement here, but it should probably
suffice for our relatively simple purposes so far.